some smaller feedback
