+++ qualification/antenna_resampled_voltage/__init__.py
@@ -1,0 +1,1 @@
+# noqa: D104
